+++ adminSiteServer/serverUtil.tsx
@@ -3,7 +3,6 @@ import * as lodash from "lodash"
 import { quote } from "shell-quote"
 import urljoin from "url-join"
 import { ADMIN_BASE_URL } from "settings"
-import * as util from "util"
 
 // Exception format that can be easily given as an API error
 export class JsonError extends Error {
@@ -39,21 +38,3 @@ export const isValidSlug = (slug: any) =>
 export const shellEscape = (str: string) => quote([str])
 
 export const absoluteUrl = (path: string) => urljoin(ADMIN_BASE_URL, path)
-
-export const splitOnLastWord = (str: string) => {
-    const endIndex = (str.lastIndexOf(" ") as number) + 1
-    return {
-        start: endIndex === 0 ? "" : str.substring(0, endIndex),
-        end: str.substring(endIndex),
-    }
-}
-
-export const execFormatted = async (
-    cmd: string,
-    args: string[],
-    verbose = true
-) => {
-    const formatCmd = util.format(cmd, ...args.map((s) => quote([s])))
-    if (verbose) console.log(formatCmd)
-    return await execWrapper(formatCmd, { silent: !verbose })
-}
