+++ site/FrontPage.tsx
@@ -11,11 +11,18 @@ import { faTwitter } from "@fortawesome/free-brands-svg-icons/faTwitter"
 import { faFacebookSquare } from "@fortawesome/free-brands-svg-icons/faFacebookSquare"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 import { faArrowRight } from "@fortawesome/free-solid-svg-icons/faArrowRight"
-import { splitOnLastWord } from "adminSiteServer/serverUtil"
 import {
     NewsletterSubscriptionForm,
     NewsletterSubscriptionContext,
 } from "site/NewsletterSubscription"
+
+const splitOnLastWord = (str: string) => {
+    const endIndex = (str.lastIndexOf(" ") as number) + 1
+    return {
+        start: endIndex === 0 ? "" : str.substring(0, endIndex),
+        end: str.substring(endIndex),
+    }
+}
 
 export const FrontPage = (props: {
     entries: CategoryWithEntries[]
